backend/requirements.txt
Flask
google-cloud-vision
pandas
openpyxl